Task 6

+++ src/app/features/courses/courses.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 import { ICourse } from './models/course.model';
 import { FilterItemsPipe } from '../../shared/pipes/filterItems.pipe';
 import { OrderByPipe } from '../../shared/pipes/orderBy.pipe';
@@ -16,6 +16,12 @@ export class CoursesComponent implements OnInit {
   coursesToShow: ICourse[] = [];
 
   searchStr: string = '';
+
+  isAddCourse: boolean = false;
+
+  onAddCourse(): void {
+    this.isAddCourse = true;
+  }
 
   constructor(
     private filterItems: FilterItemsPipe,
@@ -42,8 +48,11 @@ export class CoursesComponent implements OnInit {
   }
 
   deleteCourse($event: number): void {
-    this.coursesService.removeItem($event);
-    this.showCourses();
+    let confirmOnDelete = confirm('Do you really want to delete this course?');
+    if (confirmOnDelete) {
+      this.coursesService.removeItem($event);
+      this.showCourses();
+    }
   }
 
   searchCourses(searchString: string): void {
